backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1

sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2

pymongo<5
motor>=3.6

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

openai==1.40.0

black==24.4.2
isort==5.13.2
